krgrpc: Use *_bit() on rpc_desc_recv->flags
Closing connections will need to modify it concurrently to the desc's
thread.

+++ net/krgrpc/rpc_internal.h
@@ -50,7 +50,7 @@ struct rpc_desc_recv {
 	struct list_head list_signal_head;
 	struct rpc_desc_elem *iter;
 	struct rpc_desc_elem *iter_provided;
-	int flags;
+	unsigned long flags;		     /* bitfield */
 };
 
 enum rpc_connection_state {
